chore : ContestItem파일 이동
